requirements.txt
matplotlib==3.6.2
numpy==1.24.1
opencv_python==4.5.4.58
Pillow==9.4.0
scipy==1.10.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
tqdm==4.64.1
